split functions into files; add vectorized orient3d
