+++ include/igl/predicates/ear_clipping.cpp
@@ -8,6 +8,7 @@
 
 #include "ear_clipping.h"
 #include "predicates.h"
+#include "orient2d.h"
 #include "segment_segment_intersect.h"
 #include "../turning_number.h"
